New game state class
