+++ src/components/Scramble.vue
@@ -1,9 +1,9 @@
 <template>
-    <v-layout justify-start align-center column v-if="enabled">
+    <v-layout justify-start align-center column v-if="game.scrambling">
         <v-flex xs12 class="card-scramble">
-            <v-card v-if="working">
+            <v-card v-if="game.working">
                 <v-card-title>
-                    <h3>{{operation}}</h3>
+                    <h3>{{game.operation}}</h3>
                     <v-progress-linear :indeterminate="true"></v-progress-linear>
                 </v-card-title>
             </v-card>
@@ -11,11 +11,11 @@
                 <v-card-title class="headline">
                     <v-layout row wrap align-center>
                         <v-flex>
-                            <v-btn flat icon color="primary" @click="generateScramble">
+                            <v-btn flat icon color="primary" @click="game.generateScramble()">
                                 <v-icon>refresh</v-icon>
                             </v-btn>
                         </v-flex>
-                        <v-flex v-for="(move, index) in scramble" :key="index" :class="[{'grey--text': position > index}]">
+                        <v-flex v-for="(move, index) in game.scramble" :key="index" :class="[{'grey--text': game.position > index}]">
                             {{move}}
                         </v-flex>
                         <v-flex>
@@ -23,7 +23,7 @@
                             <v-tooltip bottom open-delay="0" v-model="tooltip" activator=".scramble_info">
                                 <span class="subheading">
                                     - <kbd>U</kbd> is <strong>white</strong> and <kbd>F</kbd> is <strong>green</strong><br/>
-                                    - <kbd>{{solveTrigger.join(' ')}}</kbd> to start solve without completing the scramble
+                                    - <kbd>{{game.solveTrigger.join(' ')}}</kbd> to start solve without completing the scramble
                                 </span>
                             </v-tooltip>
                         </v-flex>
@@ -35,110 +35,19 @@
 </template>
 
 <script lang="ts">
-import { Component, Vue } from 'vue-property-decorator'
+import { Component, Vue, Prop } from 'vue-property-decorator'
 import { EventHub, Events } from '@/classes/event-hub'
 import Timer from '@/classes/timer'
 import Worker from 'worker-loader!@/classes/cubejs.worker'
 import CubeState from '@/classes/cube-state'
-// tslint:disable-next-line:no-var-requires
-const cubejs = require('cubejs')
+import GameState from '@/classes/game-state'
 
 @Component
 export default class Scramble extends Vue {
-    private enabled = true
-    private working = true
+    @Prop()
+    private game!: GameState
+
     private tooltip = false
-
-    private operation = 'Initializing scrambler...'
-    private scramble: string[] = []
-    private position = 0
-
-    private solveTrigger = ['R', 'R\'']
-    private triggerBuffer: string[] = []
-    private lastMoveTimestamp = 0
-
-    private worker = new Worker()
-
-    private mounted() {
-        EventHub.$on(Events.cubeState, (state: Uint8Array) => this.onCubeState(state))
-        EventHub.$on(Events.cubeSolved, () => this.generateScramble())
-        EventHub.$on(Events.solveCancelled, () => this.generateScramble())
-
-        this.worker.onmessage = (event: MessageEvent) => {
-            switch (event.data.cmd) {
-                case 'init':
-                    this.generateScramble()
-                    break
-                case 'scramble':
-                    this.onScrambleGenerated(event.data.scramble)
-                    break
-            }
-        }
-
-        this.worker.postMessage({ cmd: 'init'})
-    }
-
-    private onCubeState(state: Uint8Array) {
-        if (!this.enabled || !this.scrambleAvailable() || this.scrambleCompleted()) {
-            return
-        }
-
-        const cubeState = CubeState.from(state)
-        this.checkTrigger(cubeState)
-
-        const expectedState: any = new cubejs()
-        expectedState.move(this.scramble.slice(0, this.position + 1).join(' '))
-
-        if (cubeState.visualState === expectedState.asString()) {
-            this.position++
-            if (this.scrambleCompleted()) {
-                this.onScrambleCompleted()
-            }
-        }
-    }
-
-    private onScrambleGenerated(scramble: string) {
-        this.working = false
-        this.scramble = []
-        this.position = 0
-        scramble.split(' ').forEach((s: string) => this.scramble.push(s))
-    }
-
-    private onScrambleCompleted() {
-        Timer.cubeScrambled()
-        Vue.nextTick(() => EventHub.$emit(Events.cubeScrambled, this.scramble.join(' ')))
-        this.enabled = false
-    }
-
-    private generateScramble() {
-        this.enabled = true
-        this.operation = 'Generating scramble...'
-        this.working = true
-        this.worker.postMessage({ cmd: 'scramble'})
-    }
-
-    private scrambleAvailable() {
-        return this.scramble.length > 0
-    }
-
-    private scrambleCompleted() {
-        return this.scramble.length > 0 && this.scramble.length === this.position
-    }
-
-    private checkTrigger(state: CubeState) {
-        if (state.lastmove() === this.solveTrigger[this.triggerBuffer.length] &&
-        (this.triggerBuffer.length === 0 || (Date.now() - this.lastMoveTimestamp < 200))) {
-            this.triggerBuffer.push(state.lastmove())
-            if (this.solveTrigger.join('') === this.triggerBuffer.join('')) {
-                this.triggerBuffer = []
-                this.onScrambleCompleted()
-            }
-        } else {
-            this.triggerBuffer = []
-        }
-
-        this.lastMoveTimestamp = Date.now()
-    }
 }
 </script>
